Add support for .NET Core

+++ src/Ntfs/NonResidentAttributeBuffer.cs
@@ -152,7 +152,7 @@ namespace DiscUtils.Ntfs
                     int toWrite = (int)Math.Min(remaining, _bytesPerCluster - clusterOffset);
 
                     _activeStream.ReadClusters(vcn, 1, _ioBuffer, 0);
-                    Array.Copy(buffer, offset + (focusPos - pos), _ioBuffer, clusterOffset, toWrite);
+                    Array.Copy(buffer, (int)(offset + (focusPos - pos)), _ioBuffer, (int)clusterOffset, toWrite);
                     allocatedClusters += _activeStream.WriteClusters(vcn, 1, _ioBuffer, 0);
 
                     focusPos += toWrite;
